rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faiss-cpu>=1.9.0
numpy>=2.1.0
pydantic==2.5.2
